+++ tests/integration/__init__.py
@@ -13,3 +13,4 @@ To skip integration tests (default):
 
 
 
+